requirements.txt
selenium
webdriver-manager
python-dotenv
bs4
